Move JobReportService from src to plugins
Also merge the header to the source
