--- FWCore/Services/BuildFile.xml
@@ -1,12 +1,0 @@
-<use name="FWCore/AbstractServices"/>
-<use name="FWCore/MessageLogger"/>
-<use name="FWCore/ParameterSet"/>
-<use name="FWCore/ServiceRegistry"/>
-<use name="FWCore/Utilities"/>
-<use name="boost"/>
-<use name="tinyxml2"/>
-<use name="rootcore"/>
-<use name="roothistmatrix"/>
-<export>
-  <lib name="1"/>
-</export>
