+++ FWCore/Services/test/BuildFile.xml
@@ -27,10 +27,7 @@
 
 <bin file="test_catch2_*.cc" name="testFWCoreServicesCatch2">
   <use name="catch2"/>
-  <use name="clhep"/>
   <use name="cpu_features"/>
-  <use name="FWCore/Concurrency"/>
-  <use name="FWCore/Services"/>
 </bin>
 
 <test name="TestFWCoreServicesTracerFile" command="test_tracer_file.sh"/>
